backend/requirements.txt
alembic
fastapi
passlib[bcrypt]
psycopg2-binary
pydantic
pyjwt
python-decouple
python-multipart
sqlmodel
uvicorn